zchat/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
idna==3.4
protobuf==4.24.4
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
